requirements.txt
ultralytics==8.2.90
opencv-python-headless==4.10.0.84
fastapi==0.111.1
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
streamlit==1.38.0
Pillow==10.4.0
numpy==1.26.4
requests==2.32.3
torch>=2.0.0
torchvision>=0.15.0
python-dotenv==1.0.1
